feat: Complete Milestone 1 and deploy all contracts to testnet
This commit marks the successful deployment of all four PowerGrid Network smart contracts to the POP Network testnet.

- Deployed Token, Resource Registry, Grid Service, and Governance contracts.
- Updated README with live contract addresses and improved usage guides.
- Added testnet-addresses.json for easy reference.

+++ contracts/governance/src/lib.rs
@@ -121,7 +121,7 @@ use powergrid_shared::{Proposal, ProposalType, ink_account_to_bytes};
             }
 
             let current_block = self.env().block_number();
-            let voting_end = current_block.saturating_add(self.voting_duration_blocks as u32);
+            let voting_end = (current_block as u64).saturating_add(self.voting_duration_blocks);
             let proposal_id = self.next_proposal_id;
 
             let proposal = Proposal {
